Do a better job finding the datepicker
If the datepicker was instantiated from inside a `within` block, finding
the datepicker would not work because it could not be found in the
page's current scope. Adding our own temporary `within` block can break
out of the current scope and again look in the entire document.

This also scopes the datepicker find to `div`s that are direct children
of the body, to help avoid any confusion with other elements with the
datepicker class (like inputs).